Move snapshot creator to protocol
