+++ pkg/protocol/snapshotcreator/options.go
