+++ pkg/protocol/snapshotcreator/snapshotcreator.go
